[RFC]rdb: async free old key/value when bgsave is in progress
On key overwrite the parent process would:
  1) find the old dict entry
  2) set the new value for the old dict entry
  3) free the old value

In step 2 and 3 parent will need to modify the memory pages where
the old dict entry/value is on. In a snapshot context this would
bring lots of memory CoW in the parent process if the overwrite speed
is high.

This patch introduced a new mechanism to reduce the CoW on key overwrites
in a snapshot context. On key overwrite:
  1) pull the old dict entry out of the dict table
  2) alloc a new dict entry for the old key
  3) set the value for the new dict entry
  4) create a bio job(LAZY_FREE_ENTRY) to free the old key/value in background
     thread
  5) in the background thread, the free wil be defered if we detect snapshot is
     in progress

The current approach is quite naive and each instance may tempeory hold bigger
memory size it should have. But we could do more tunings on the async free speed
to control the CoW size vs RSS memory size.

Signed-off-by: Yuan Zhou <yuan.zhou@intel.com>